gets filesize, 64bit/32bit division fails still

+++ btp.h
@@ -10,14 +10,14 @@ typedef struct _BTP_CHUNK_HEADER {
 } BTP_CHUNK_HEADER;
 
 typedef struct _BTP_FILE_HEADER {
-    long long size; // size of the file 64-bit
+    off_t size; // size of the file 64-bit
     FILE* hFile; // handle to the file
-    long long nChunks; // number of chunks in the file 64-bit
-    long long ib1; // index of the chunk loaded in the first buffer
+    off_t nChunks; // number of chunks in the file 64-bit
+    off_t ib1; // index of the chunk loaded in the first buffer
     BTP_CHUNK_HEADER* pb1; // chunk header for the first buffer
-    long long ib2; // index of the chunk loaded in the second buffer
+    off_t ib2; // index of the chunk loaded in the second buffer
     BTP_CHUNK_HEADER* pb2; // chunk header for the second buffer
-    long long ibc; // index of the chunk loaded in the check buffer
+    off_t ibc; // index of the chunk loaded in the check buffer
     BTP_CHUNK_HEADER* pbc; // chunk header for the check buffer
 } BTP_FILE_HEADER;
